Fix packaging exports, docs import paths, and CI packaging gates
- Split the root export's types per import/require condition and emit
  dist/index.d.cts so CJS consumers no longer get ESM-flavored types
  (attw 'Masquerading as ESM').
- Fix the tailwind preset for "type": "module": the CommonJS
  implementation now lives in tailwind-preset.cjs and tailwind-preset.js
  is a real ESM wrapper (importing the preset previously threw
  'module is not defined' under strict Node ESM resolution). Add type
  declarations for the preset entrypoint.
- Export AIProvider as a value from the root so the documented
  'extends AIProvider' pattern works; it was previously type-only.
- Update README and docs to import from the package root instead of the
  nonexistent /providers and /storage subpaths, which fail to resolve.
- Make publint (--strict) and are-the-types-wrong blocking in CI: the
  step previously ran with continue-on-error and also called a
  nonexistent 'are-the-types-wrong' binary (the CLI is 'attw'), so it
  always failed silently. Add the same packaging validation to the
  publish workflow, and remove the '|| true' escape hatches from the
  compatibility job.
- Tighten react/react-dom peer ranges to ^19.0.0 to match what the
  library is developed and tested against.
