+++ package.json
@@ -41,20 +41,33 @@
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
     },
     "./styles.css": "./dist/styles.css",
     "./tailwind-preset": {
-      "import": "./tailwind-preset.js",
-      "require": "./tailwind-preset.cjs"
+      "import": {
+        "types": "./tailwind-preset.d.ts",
+        "default": "./tailwind-preset.js"
+      },
+      "require": {
+        "types": "./tailwind-preset.d.cts",
+        "default": "./tailwind-preset.cjs"
+      }
     }
   },
   "files": [
     "dist",
     "tailwind-preset.js",
-    "tailwind-preset.cjs"
+    "tailwind-preset.cjs",
+    "tailwind-preset.d.ts",
+    "tailwind-preset.d.cts"
   ],
   "sideEffects": [
     "**/*.css",
@@ -62,7 +75,7 @@
   ],
   "scripts": {
     "dev": "vite build --watch",
-    "build": "tsc -b && vite build",
+    "build": "tsc -b && vite build && node -e \"require('fs').copyFileSync('dist/index.d.ts','dist/index.d.cts')\"",
     "lint": "eslint .",
     "lint:fix": "tsc --noEmit && eslint . --fix",
     "typecheck": "tsc --noEmit",
@@ -96,8 +109,8 @@
     "@ai-sdk/openai-compatible": ">=1.0.0",
     "@openrouter/ai-sdk-provider": ">=1.0.0",
     "ollama-ai-provider-v2": ">=1.0.0",
-    "react": ">=16.8.0",
-    "react-dom": ">=16.8.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-hook-form": "^7.62.0"
   },
   "peerDependenciesMeta": {
