+++ tailwind-preset.d.cts
@@ -1,0 +1,4 @@
+import type { Config } from 'tailwindcss';
+
+declare const preset: Partial<Config>;
+export = preset;
